Move DevicePluginStub implementation into new plugin package
