+++ pkg/kubelet/cm/devicemanager/plugin/v1beta1/stub.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package devicemanager
+package v1beta1
 
 import (
 	"context"
