+++ test/images/sample-device-plugin/VERSION
@@ -1,1 +1,1 @@
-1.3
+1.4
